+++ constraint-solver/src/solver/base.rs
@@ -1,5 +1,4 @@
 use itertools::Itertools;
-use num_traits::Zero;
 use powdr_number::{ExpressionConvertible, FieldElement};
 
 use crate::constraint_system::{
@@ -157,9 +156,9 @@ where
 
         let constraints = constraints
             .into_iter()
-            .filter(|c| !c.is_zero())
+            .filter(|c| !c.is_redundant())
             .flat_map(|constr| {
-                self.try_extract_boolean(&constr)
+                self.try_extract_boolean(constr.as_ref())
                     .into_iter()
                     .chain(std::iter::once(constr))
             })
@@ -170,7 +169,7 @@ where
             .collect_vec();
 
         self.constraint_system
-            .add_algebraic_constraints(constraints.into_iter().filter(|c| !c.is_zero()));
+            .add_algebraic_constraints(constraints.into_iter().filter(|c| !c.is_redundant()));
     }
 
     fn add_bus_interactions(
@@ -272,7 +271,7 @@ where
     /// by introducing new boolean variables.
     fn try_extract_boolean(
         &mut self,
-        constr: &AlgebraicConstraint<GroupedExpression<T, V>>,
+        constr: AlgebraicConstraint<&GroupedExpression<T, V>>,
     ) -> Option<AlgebraicConstraint<GroupedExpression<T, V>>> {
         let result = self
             .boolean_extractor
@@ -293,7 +292,7 @@ where
         constr: AlgebraicConstraint<GroupedExpression<T, V>>,
     ) -> impl Iterator<Item = AlgebraicConstraint<GroupedExpression<T, V>>> {
         let mut constrs = vec![constr.clone()];
-        if !constr.is_affine() {
+        if !constr.expression.is_affine() {
             let linearized = self.linearizer.linearize_expression(
                 constr.expression,
                 &mut || self.var_dispenser.next_linear(),
@@ -365,12 +364,12 @@ where
         while let Some(item) = self.constraint_system.pop_front() {
             let effects = match item {
                 ConstraintRef::AlgebraicConstraint(c) => {
-                    if let Some((v1, expr)) = try_to_simple_equivalence(c) {
+                    if let Some((v1, expr)) = try_to_simple_equivalence(c.clone()) {
                         self.apply_assignment(&v1, &expr);
                         continue;
                     }
                     c.solve(&self.range_constraints)
-                        .map_err(Error::QseSolvingError)?
+                        .map_err(Error::AlgebraicSolverError)?
                         .effects
                 }
                 ConstraintRef::BusInteraction(b) => b
@@ -568,7 +567,7 @@ where
 ///
 /// Note: Does not find all cases of equivalence.
 fn try_to_simple_equivalence<T: RuntimeConstant, V: Clone + Ord + Eq>(
-    constr: &AlgebraicConstraint<GroupedExpression<T, V>>,
+    constr: AlgebraicConstraint<&GroupedExpression<T, V>>,
 ) -> Option<(V, GroupedExpression<T, V>)> {
     if !constr.expression.is_affine() {
         return None;
