+++ constraint-solver/src/solver/exhaustive_search.rs
@@ -155,11 +155,12 @@ where
         .constraints_referencing_variables(assignments.keys().cloned())
         .map(|constraint| match constraint {
             ConstraintRef::AlgebraicConstraint(identity) => {
-                let mut identity = identity.clone();
+                let mut identity = identity.cloned();
                 for (variable, value) in assignments.iter() {
                     identity.substitute_by_known(variable, &T::from(*value));
                 }
                 identity
+                    .as_ref()
                     .solve(range_constraints)
                     .map(|result| result.effects)
                     .map_err(|_| ContradictingConstraintError)
